+++ packages/core/src/editor/actions/FreCreatePartCommand.ts
@@ -1,4 +1,3 @@
-import { FreLogger } from "../../logging/index";
 import { BTREE } from "../../util/index";
 import { EMPTY_POST_ACTION, FrePostAction, ReferenceShortcut } from "./FreAction";
 import { Box } from "../boxes";
@@ -6,9 +5,8 @@ import { isString, FreTriggerUse, triggerTypeToString } from "./FreTriggers";
 import { FreEditor } from "../FreEditor";
 import { FreBinaryExpression, FreNode } from "../../ast";
 import { FreLanguage } from "../../language";
-import { FreCommand } from "./FreCommand";
+import { FreCommand, FRECOMMAND_LOGGER } from "./FreCommand";
 
-const LOGGER = new FreLogger("FreCommand");
 /**
  * Command to create a part (child) of a FreElement.
  * The FreElement of the box on which this command is executed shpould have a property with `propertyName` of
@@ -24,7 +22,7 @@ export class FreCreatePartCommand extends FreCommand {
         this.propertyName = propertyName;
         this.conceptName = conceptName;
         this.referenceShortcut = referenceShortcut;
-        LOGGER.log("+++++++++++++++ Create part command " + propertyName + ", " + conceptName);
+        FRECOMMAND_LOGGER.log("+++++++++++++++ Create part command " + propertyName + ", " + conceptName);
     }
 
     /**
@@ -36,7 +34,7 @@ export class FreCreatePartCommand extends FreCommand {
      */
     execute(box: Box, trigger: FreTriggerUse, editor: FreEditor, index: number): FrePostAction {
         // todo make index optional and set the default value to -1;
-        LOGGER.log(
+        FRECOMMAND_LOGGER.log(
             "CreatePartCommand: trigger [" +
                 triggerTypeToString(trigger) +
                 "] part: " +
@@ -58,7 +56,7 @@ export class FreCreatePartCommand extends FreCommand {
             console.error("ActionBox action: Unexpected new element undefined");
             return EMPTY_POST_ACTION;
         }
-        LOGGER.log(
+        FRECOMMAND_LOGGER.log(
             `FreCreatePartCommand: setting/adding to ${propName} of ${box.node.freId()} (${box.node.freLanguageConcept()}) to ${newElement.freId()} (${newElement.freLanguageConcept()})`,
         );
         if (FreLanguage.getInstance().classifierProperty(ownerConcept, propName).isList) {
@@ -83,7 +81,7 @@ export class FreCreatePartCommand extends FreCommand {
         return function () {
             // editor.selectElement(newElement);
             // tslint:disable-next-line:max-line-length
-            LOGGER.log(
+            FRECOMMAND_LOGGER.log(
                 "CreatePartCommand: newElement:" +
                     newElement.freId() +
                     " " +
